[APIS-1106] Break LB package cycles and group the metrics classes
