+++ src/jdbc/cubrid/jdbc/lb/RoutingMetricsRecorder.java
@@ -31,18 +31,18 @@
 package cubrid.jdbc.lb;
 
 import cubrid.jdbc.lb.config.Endpoint;
+import cubrid.jdbc.lb.config.LbLogConfig;
 import cubrid.jdbc.lb.config.LoadBalanceSettings;
 import cubrid.jdbc.lb.failover.PhysicalRecoveryContext;
 import cubrid.jdbc.lb.failover.PhysicalRecoveryResult;
 import cubrid.jdbc.lb.log.LbLog;
-import cubrid.jdbc.lb.log.LbLogConfig;
+import cubrid.jdbc.lb.metrics.RuntimeMetrics;
 import cubrid.jdbc.lb.route.Router;
 import cubrid.jdbc.lb.sql.HintParser;
 import cubrid.jdbc.lb.sql.HintSet;
 import cubrid.jdbc.lb.sql.SqlClassification;
 import cubrid.jdbc.lb.sql.SqlClassifier;
 import cubrid.jdbc.lb.sql.SqlLiteralMasker;
-import cubrid.jdbc.lb.state.RuntimeMetrics;
 import cubrid.jdbc.lb.state.SessionRoutingState;
 import java.util.logging.Level;
 import java.util.logging.Logger;
