+++ src/test/cubrid/jdbc/lb/it/LbVendorApiInheritanceIT.java
@@ -49,7 +49,7 @@ import cubrid.jdbc.jci.UConnection;
 import cubrid.jdbc.lb.LoadBalanceConnection;
 import cubrid.jdbc.lb.SessionLeg;
 import cubrid.jdbc.lb.config.Endpoint;
-import cubrid.jdbc.lb.state.RuntimeMetrics;
+import cubrid.jdbc.lb.metrics.RuntimeMetrics;
 import cubrid.sql.CUBRIDOID;
 import cubrid.sql.CUBRIDOIDImpl;
 import java.sql.Connection;
