+++ src/test/cubrid/jdbc/lb/log/LbShutdownRecordsTest.java
@@ -33,12 +33,14 @@ package cubrid.jdbc.lb.log;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import cubrid.jdbc.lb.config.LbLogConfig;
 import cubrid.jdbc.lb.config.LoadBalanceSettings;
 import cubrid.jdbc.lb.config.LoadBalanceUrlParser;
+import cubrid.jdbc.lb.metrics.BindingView;
+import cubrid.jdbc.lb.metrics.MetricsRegistry;
+import cubrid.jdbc.lb.metrics.RuntimeMetrics;
+import cubrid.jdbc.lb.report.LbDistLog;
 import cubrid.jdbc.lb.sql.SqlClassification;
-import cubrid.jdbc.lb.state.BindingView;
-import cubrid.jdbc.lb.state.MetricsRegistry;
-import cubrid.jdbc.lb.state.RuntimeMetrics;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
