+++ src/test/cubrid/jdbc/lb/metrics/MetricsExportersTest.java
@@ -31,16 +31,9 @@
 package cubrid.jdbc.lb.metrics;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import cubrid.jdbc.jci.UUnreachableHostList;
-import cubrid.jdbc.lb.state.BindingView;
-import cubrid.jdbc.lb.state.MetricsRegistry;
-import cubrid.jdbc.lb.state.RuntimeMetrics;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.After;
@@ -116,83 +109,6 @@ public class MetricsExportersTest {
 
         assertEquals(1, MetricsRegistry.countRoOnRwSessions());
         assertTrue(MetricsExporters.renderPrometheus().contains("lb_ro_on_rw_sessions 1"));
-    }
-
-    @Test
-    public void rotationKeepsAtMostMaxFilesGenerationsAndDropsTheOldest() throws IOException {
-        File dir = new File(System.getProperty("java.io.tmpdir"), "lb-metrics-rotate-test");
-        deleteRecursively(dir);
-        assertTrue(dir.mkdirs());
-        File live = new File(dir, "lb-metrics.csv");
-
-        // 3 rotations with maxFiles=2: the live file becomes .1, the old .1 becomes .2, and the
-        // generation that would have become .3 is deleted.
-        for (int round = 1; round <= 3; round++) {
-            write(live, "round" + round);
-            MetricsExporters.rotateIfOversized(live, 1L, 2);
-        }
-
-        assertFalse("live file is rotated away; the next write re-creates it", live.exists());
-        assertEquals("round3", read(new File(dir, "lb-metrics.csv.1")));
-        assertEquals("round2", read(new File(dir, "lb-metrics.csv.2")));
-        assertFalse(
-                "generation beyond maxFiles must be deleted",
-                new File(dir, "lb-metrics.csv.3").exists());
-        deleteRecursively(dir);
-    }
-
-    @Test
-    public void rotationIsSkippedBelowTheSizeLimitAndWhenDisabled() throws IOException {
-        File dir = new File(System.getProperty("java.io.tmpdir"), "lb-metrics-rotate-skip-test");
-        deleteRecursively(dir);
-        assertTrue(dir.mkdirs());
-        File live = new File(dir, "lb-metrics.csv");
-        write(live, "keep");
-
-        MetricsExporters.rotateIfOversized(live, 1024L * 1024L, 5); // under the limit
-        assertTrue(live.exists());
-
-        MetricsExporters.rotateIfOversized(live, 0L, 5); // rotation disabled
-        assertTrue(live.exists());
-        assertEquals("keep", read(live));
-        deleteRecursively(dir);
-    }
-
-    private static void write(final File f, final String body) throws IOException {
-        FileOutputStream out = new FileOutputStream(f, false);
-        try {
-            out.write(body.getBytes("UTF-8"));
-        } finally {
-            out.close();
-        }
-    }
-
-    private static String read(final File f) throws IOException {
-        byte[] buf = new byte[(int) f.length()];
-        java.io.InputStream in = new java.io.FileInputStream(f);
-        try {
-            int off = 0;
-            while (off < buf.length) {
-                int n = in.read(buf, off, buf.length - off);
-                if (n < 0) {
-                    break;
-                }
-                off += n;
-            }
-        } finally {
-            in.close();
-        }
-        return new String(buf, "UTF-8");
-    }
-
-    private static void deleteRecursively(final File dir) {
-        File[] kids = dir.listFiles();
-        if (kids != null) {
-            for (int i = 0; i < kids.length; i++) {
-                kids[i].delete();
-            }
-        }
-        dir.delete();
     }
 
     private static void register(
