+++ src/test/cubrid/jdbc/lb/metrics/MetricsRegistryConcurrentFoldTest.java
@@ -28,7 +28,7 @@
  *
  */
 
-package cubrid.jdbc.lb.state;
+package cubrid.jdbc.lb.metrics;
 
 import static org.junit.Assert.assertEquals;
